Move ellipsis tooltip to misc
https://coveord.atlassian.net/browse/JSUI-2406
