+++ src/ui/Misc/TextEllipsisTooltip.ts
